Time: 8 ms (90.10%), Space: 9.7 MB (74.27%) - LeetHub

+++ 1239-maximum-length-of-a-concatenated-string-with-unique-characters/1239-maximum-length-of-a-concatenated-string-with-unique-characters.cpp
@@ -23,6 +23,6 @@ public:
                 }
             }
         }
-        return (int) maxLen;
+        return maxLen;
     }
 };
